\name{covglasso-package}
\alias{covglasso-package}
\docType{package}

\title{Sparse covariance matrix estimation}

\description{Fast and direct estimation of a sparse covariance matrix via covariance graphical lasso and coordinate descent algorithm.}

\details{A package implementing direct estimation of a sparse covariance matrix corresponding to a Gaussian covariance graphical model. Estimation is performed by solving the covariance graphical lasso using a fast coordinate descent algorithm.}

\section{How to cite this package}{
To cite \pkg{covglasso} in publications use:
\preformatted{
Fop, M. (2020). covglasso: Sparse Covariance Matrix Estimation, R package version 1.0,
https://CRAN.R-project.org/package=covglasso
}}

\author{
Michael Fop.

Maintainer: Michael Fop \email{michael.fop@ucd.ie}
}

\references{
Bien, J., Tibshirani, R.J. (2011). Sparse estimation of a covariance matrix. \emph{Biometrika}, 98(4), 807–820.

Wang, H. (2014). Coordinate descent algorithm for covariance graphical lasso. \emph{Statistics and Computing}, 24:521.
}

\keyword{package}
